+++ inc/flagship-library/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog for Flagship Library
 
+## 1.4.3
+
+Made a few minor adjustments to the TinyMCE button names and improved the default author box templates. Also introduced a function to output a link to the parent theme's landing page.
+
+## 1.4.2
+
+- Fixed a bug which was causing author box settings to be overwritten in the admin.
+
 ## 1.4.1
 
 - Fixed a fatal error by making Flagship_Library::$dir public rather than protected.